fix: move integration test utils to separate file
